(dev) replaced empty funcs to main.py
